+++ greedy.hpp
@@ -16,12 +16,12 @@ struct CompareNodes {
 };
 
 //selects the node coloration with the lowest maxValue in this step. returns the selected color if coloring was possible and UNDEFINED if it was not. updates the avaliableColors matrix
-int getNodeBestStepColor(int colorCount, double* totalColorWeights, int node, int* avaliableColors);
+int getNodeBestStepColor(int colorCount, float* totalColorWeights, int node, int* avaliableColors);
 
 //selects the node coloration with the lowest number of conflicts. if there is more than one coloration with the lowest number of conflicts, the one with lowest total weights is selected
-int getNodeRecolorationColor(int colorCount, int** nodeAdjacencyList, int* coloration, double* totalColorWeights, int node, int adjacentNodeQuantity, int* recolorationCount);
+int getNodeRecolorationColor(int colorCount, int** nodeAdjacencyList, int* coloration, float* totalColorWeights, int node, int adjacentNodeQuantity, int* recolorationCount);
 
 //colors the graph following the getNodeBestStepColor() heuristic. returns UNDEFINED if was not able to construct following the heuristic
-int greedyConstruction(int nodeCount, int colorCount, int** nodeAjacencyList, double* weights, int* coloration, double* totalColorWeights, int* adjacentNodeQuantity, int* avaliableColors, clock_t t0);
+int greedyConstruction(int nodeCount, int colorCount, int** nodeAjacencyList, float* weights, int* coloration, float* totalColorWeights, int* adjacentNodeQuantity, int* avaliableColors, clock_t t0);
 
 #endif /* _GREEDY_H_ */
