+++ main.cpp
@@ -6,13 +6,13 @@ int main() {
     int nodeCount = 0;
     int edgeCount = 0;
     int colorCount = 0;
-    double minimal_possible_max_value = numeric_limits<double>::infinity()*-1;
+    float minimal_possible_max_value = numeric_limits<float>::infinity()*-1;
 
     scanf("%d", &nodeCount);
     scanf("%d", &edgeCount);
     scanf("%d", &colorCount);
 
-    double weights[nodeCount];
+    float weights[nodeCount];
 
     //reading weights
     for(int i = 0; i < nodeCount; i++){
@@ -90,8 +90,8 @@ int main() {
 
     bool timeout = false;
 
-    double maxValue = numeric_limits<double>::infinity();
-    double newMaxValue;
+    float maxValue = numeric_limits<float>::infinity();
+    float newMaxValue;
 
     srand(time(NULL));
 
@@ -104,7 +104,7 @@ int main() {
         }
 
         t_diff = clock() - t0;
-        double time_taken = ((double)t_diff)/CLOCKS_PER_SEC;
+        float time_taken = ((float)t_diff)/CLOCKS_PER_SEC;
 
         cout << "time taken for GRASP iteration in seconds: " << time_taken << "\n";
 
@@ -113,7 +113,7 @@ int main() {
             break;
         }
 
-        double newMaxValue = grasp(nodeCount, colorCount, nodeAjacencyList, weights, currentColorationPointer, adjacentNodeListLength, &adjacencyHash, t0, &edges[0][0], edgeCount);
+        float newMaxValue = grasp(nodeCount, colorCount, nodeAjacencyList, weights, currentColorationPointer, adjacentNodeListLength, &adjacencyHash, t0, &edges[0][0], edgeCount);
 
         //check if solution is valid
